Ajuste de primera fase.
*Aislamiento de inicialización de parámetros.
*Inicialización más eficiente del objeto "op" perteneciente a la clase Operacion.
*Uso de función let.
*Inclusión de cadena previamente declarada en el fichero String.xml

+++ PracticaUno/app/src/main/java/com/practicauno/Operation.kt
@@ -4,5 +4,5 @@ class Operation() {
     fun additionOp(numberOne: Int, numberTwo: Int): Int = numberOne + numberTwo
     fun subtractionOp(numberOne: Int, numberTwo: Int): Int = numberOne - numberTwo
     fun multiplicationOp(numberOne: Int, numberTwo: Int): Int = numberOne * numberTwo
-    fun divisionOp(numberOne: Int, numberTwo: Int): Float = (numberOne / numberTwo).toFloat()
+    fun divisionOp(numberOne: Int, numberTwo: Int): Double = numberOne.toDouble() / numberTwo.toDouble()
 }
